review delete complete

+++ src/components/MyReviews/ReviewTable.js
@@ -1,17 +1,46 @@
 import React from "react";
 import Button from "react-bootstrap/Button";
-
+import Swal from "sweetalert2";
 
 const ReviewTable = ({ review }) => {
   const { serviceName, reviewTxt, _id } = review;
 
   const handleDelete = (_id) => {
-    fetch(`http://localhost:5000/delete/${_id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: "btn btn-success",
+        cancelButton: "btn btn-danger",
+      },
+      buttonsStyling: false,
+    });
+    swalWithBootstrapButtons
+      .fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          fetch(`http://localhost:5000/delete/${_id}`, {
+            method: "DELETE",
+          })
+            .then((res) => res.json())
+            .then((data) => {
+              if (data.data.acknowledged) {
+                swalWithBootstrapButtons.fire(
+                  "Deleted!",
+                  "Review has been deleted.",
+                  "success"
+                );
+              } else {
+                console.log(data);
+              }
+            });
+        }
       });
   };
   return (
@@ -25,7 +54,7 @@ const ReviewTable = ({ review }) => {
         <Button className="me-2 " variant="primary">
           Edit
         </Button>
-        <Button onClick={handleDelete} variant="danger">
+        <Button onClick={() => handleDelete(_id)} variant="danger">
           Delete
         </Button>
       </td>
